Refactor and make csv
